requirements.txt
chainlit==0.7.700
langchain_core==0.2.21
langchain==0.2.9
langchain_community==0.2.7
langgraph==0.1.8
beautifulsoup4==4.12.3 
duckduckgo_search==6.2.1
praw==7.7.1
python-dotenv==1.0.1